fixed various test which struggle with mixed prepared statements
