+++ core/src/main/java/org/polypheny/db/util/Util.java
@@ -89,6 +89,7 @@ import java.util.RandomAccess;
 import java.util.Set;
 import java.util.StringTokenizer;
 import java.util.TimeZone;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.regex.Matcher;
@@ -100,6 +101,8 @@ import org.apache.calcite.avatica.util.Spaces;
 import org.apache.calcite.linq4j.Ord;
 import org.polypheny.db.algebra.AlgCollation;
 import org.polypheny.db.algebra.AlgFieldCollation;
+import org.polypheny.db.algebra.AlgNode;
+import org.polypheny.db.algebra.AlgVisitor;
 import org.polypheny.db.algebra.constant.Kind;
 import org.polypheny.db.algebra.constant.Monotonicity;
 import org.polypheny.db.algebra.fun.AggFunction;
@@ -2051,6 +2054,24 @@ public class Util {
 
     public static Collation getDefaultCollation() {
         return Collation.COERCIBLE;
+    }
+
+
+    public static boolean containsEntity( AlgNode a ) {
+        final AtomicBoolean contains = new AtomicBoolean( false );
+
+        AlgVisitor visitor = new AlgVisitor() {
+
+            @Override
+            public void visit( AlgNode node, int ordinal, AlgNode parent ) {
+                if ( node.getEntity() != null ) {
+                    contains.set( true );
+                }
+                super.visit( node, ordinal, parent );
+            }
+        };
+        visitor.go( a );
+        return contains.get();
     }
